faang style rag upload
